requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
torch>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
pytest>=7.4.0
